Update finding-data-with-views.markdown
I'm pretty sure there are some unbalanced double quotes and curly braces in the code samples for design documents with map and reduce functions. As for the viewmode system property, having to set viewmode to development to set view mode to production doesn't make a lot of sense to me. So I'm guessing the problem is in the documentation but it may really be in the implementation. I don't know yet as I haven't tested that part of the SDK.